updated gameView prompt player bet to double

+++ src/controller/Controller.java
@@ -47,9 +47,9 @@ public class Controller implements ControllerInterface {
         for (Player p: this.myTable.getPlayers()) {
             int playerHash = p.getID();
             this.myGameView.updateMainPlayer(playerHash);
-            int min = this.myTable.getTableMin();
-            int max = (int) Math.min(this.myTable.getTableMax(), p.getBankroll());
-            System.out.printf("min: %d, max: %d\n", min, max);
+            double min = this.myTable.getTableMin();
+            double max = Math.min(this.myTable.getTableMax(), p.getBankroll());
+            System.out.printf("min: %.1f, max: %.1f\n", min, max);
             double wager = this.myGameView.promptPlayerBet(min, max);
             // double wager = 10;
             int betID = this.myTable.placeEntryBet(playerHash, this.myEntryBet, wager);
